foundations/engine/requirements.txt
pystemmer
